15.2. Doctrine postLoad Listener
Creating the Entity Listener
To set this field, we already know two solutions: a custom data provider or an event listener that sets the field early in the request.
But neither of those options will set this value everywhere. The data provider is only used for API requests and the event subscriber won't set the field inside the CLI - like when running a command.
And maybe that's okay! But this last solution will work everywhere: a Doctrine "postLoad" listener: a function that is called each time an entity is loaded from the database.
The only downside to this solution is that the postLoad listener is always called when you query for an object... even if you'll never use the isMvp field. Though... I'll show you a trick to get around this.
To create a Doctrine listener, well, you can technically create a Doctrine event subscriber, listener or something called an entity listener... and they all basically work the same.
In part 2 of this series, we created an entity listener... and I kind of like those. In src/Doctrine, we created this entity listener to automatically set the owner field to the currently authenticated User for cheese listings.
Anyways, create a new PHP class and call it UserSetIsMvpListener.
Listeners don't extend anything: the only rule is that you need to have a public function with the name of the event that you want to listen to. So postLoad() with a User $user argument since we'll hook this up as an entity listener for that class.
We'll do that in a minute.
First, let's finish the logic: $user->setIsMvp() with strpos() to look inside of $user->getUsername() for the string cheese. If this is not false, then this user is an MVP.

LazyString for Lazy Computation
Since this method will be called every time any User object is loaded from the database, we need to make sure that any logic we run here is fast, especially because the calculated field might not even be needed in all situations! So... if your logic isn't super fast... does that mean you can't use a postLoad listener?
Not necessarily. Another option is to create a LazyString. That's... literally a class inside of Symfony that helps you create strings that won't be evaluated until later - if and when they are needed. Try LazyString::fromCallable() as a neat way to make this custom field lazy.
Registering the Entity Listener
Anyways, entity listeners are not one of the situations where Symfony will magically find our class and start calling the postLoad() method. Nope, we need to tell Doctrine about this with some config. First, inside User, all the way on top, add @ORM\EntityListeners() and pass this an array with our listener class UserSetIsMvpListener::class
I also need to add the use statement manually because PhpStorm doesn't auto-complete it. Add use App\Doctrine\UserSetIsMvpListener
As soon as I do that, I can hold Cmd or Ctrl and PhpStorm does recognize the class.

Finally, we need a tiny bit of service configuration to get this working, which is kind of annoying but necessary. Open config/services.yaml... copy the entity listener from the last tutorial, paste and change the name to UserSetIsMvpListener.
Phew! With any luck, that should be enough for our function to be called every time a user is loaded from the database. Try the test:

  symfony php bin/phpunit --filter=testGetUser
We got it! So that is yet another way to populate a custom field when your API resource is an entity class.
But what if you need an API resource that is completely custom: like its data comes from a bunch of different database tables or... it comes from somewhere else entirely? Yep, it's time to create a 100% custom, non-entity API resource.