16.2. Completely Custom Resource
But we could have also solved this "Ryan hates underscores" problem in a more global way. Search for "API Platform operation path naming" to find a spot on their docs that talks about how the URLs are generated.
Whenever you have an entity - like DailyStats - there is a process inside API Platform that converts the word DailyStats to daily_stats or daily-stats. That's called the path_segment_name_generator and it's something you can customize. It uses underscores by default, but there's a pre-built service that loves dashes.
Copy the example config. Then, remove the shortName option and in api_platform.yaml paste that config anywhere

 22 lines  config/packages/api_platform.yaml
api_platform:
    path_segment_name_generator: api_platform.path_segment_name_generator.dash
Back on our docs, let's try it! Beautiful! The URL is still /api/daily-stats.

Next: let's make these operations actually work, starting with the get collection operation. How do we do that? By adding a data provider... and a few other details that are special when you working with a class that is not an entity.