17.3. Custom Resource Data Provider
You Need a "get" item Operation
The problem in this situation is that, inside DailyStats, we've specifically said that we don't want any item operations
This kind of confuses API Platform, which has a brief existential crises before saying:

  How can I generate a URL to the item operation if there aren't any! Ah!

The easiest solution is to add the get itemOperation... and we are going to do that later. But let's pretend that we don't want to use that workaround because we don't want an item operation.
The solution is... well... it's still an ugly workaround. I'll paste in some config... and then I need to add a use statement for this NotFoundAction class.
So... this basically says:

  I do want a get item operation... but if anybody goes to it, I want to execute the NotFoundAction which will rudely return a 404 response.

So there is an item operation... but not really.
Anyways, if we try it again... our next error!

  No identifiers defined for resource DailyStats.

This is much more understandable. Every resource needs to have a unique identifier, which is used to generate the IRI. Usually this is an ID or UUID, which we'll talk about later in the tutorial.