17.4. Custom Resource Data Provider
Adding the identifier
To give our DailyStats an identifier, we could add a public $id or public $uuid... but we don't need to! Why? Because the $date is already a unique identifier: we're only going to have one DailyStats per day
How do we tell API Platform to use this property as the identifier? In Doctrine, it happens automatically thanks to the @ORM\Id annotation
When you're not in an entity - or if you want to use a field other than the database id in an entity - you can add @ApiProperty() with identifier=true
Cool! Let's refresh and celebrate! With... another error:

  DateTime could not be converted to string

Oh, duh! API Platform needs to convert our identifier into a string. Since this is a Datetime object... that doesn't work.