18.1. Property Metadata Thanks to the data collection provider, our endpoint returns one result... but there are no fields! Why not? Normally if you don't set a normalizationContext - like we did in User with normalizationContext and groups:
 288 lines  src/Entity/User.php
/**
 * @ApiResource(
 *     normalizationContext={"groups"={"user:read"}},
 * )
 */
class User implements UserInterface
{
}
Then your object will be serialized with no serialization groups... which basically means that every property will be included.

But... we are not see that at all! This is due to something we did in a previous tutorial. In src/Serializer/AdminGroupsContextBuilder.php, we added code to give you extra groups if you're an admin:

 37 lines  src/Serializer/AdminGroupsContextBuilder.php
final class AdminGroupsContextBuilder implements SerializerContextBuilderInterface
{
    public function createFromRequest(Request $request, bool $normalization, ?array $extractedAttributes = null): array
    {
        $context = $this->decorated->createFromRequest($request, $normalization, $extractedAttributes);
        $context['groups'] = $context['groups'] ?? [];
        $isAdmin = $this->authorizationChecker->isGranted('ROLE_ADMIN');
        if ($isAdmin) {
            $context['groups'][] = $normalization ? 'admin:read' : 'admin:write';
        }
        $context['groups'] = array_unique($context['groups']);
        return ;
    }
}
But to do this, if the groups are not set on the $context, we initialized them to an empty array:

 37 lines  src/Serializer/AdminGroupsContextBuilder.php
final class AdminGroupsContextBuilder implements SerializerContextBuilderInterface
{
    public function createFromRequest(Request $request, bool \normalization, ?array $extractedAttributes = null): array
    {
        $context['groups'] = $context['groups'] ?? [];
    }
}
Thanks to this, if we don't have a normalization group on the resource, instead of serializing everything, it will serialize nothing because it thinks we want to serialize no groups. It's... a quirk in our project.