19.1. Why/When a Many Relation is IRI Strings vs Embedded
I want to show you something a bit surprising about how our custom API resource is being serialized. To make things more realistic, add a $stats2, and let's actually use 1000 visitors for day one - our site is at least kind of popular - 2000 for $stats2 and set this to minus 1 days from now. Put that new $stats2 into the return statement
Both of these DailyStats are missing one thing: we're passing an empty array as the fourth argument, which is supposed to be a collection of the most popular CheeseListing objects. I don't really care about the "most popular" part, so let's just pass 5 random listings.
Create a public function __construct() and autowire CheeseListingRepository. I'll hit Alt+Enter and go to "Initialize properties" to create that property and set it
Before the stats, fetch some listings with $listings = $this->cheeseListingRepository->findBy() and pass an empty array of criteria, an empty array of "order by" and 5 to get a max of 5 listings
Use this to replace the empty array: $listings and $listings

Embedded Serialization Versus Inline IRI
Easy enough! Let's see what it looks like! Back at the browser, refresh and... no surprise! It works! Wait... it looks... weird. Each listing has @id and @type... but no other fields. And... hmm. That actually makes sense.
In DailyStats, to serialize - or "normalize" - we're using a group called daily-stats:read:
 61 lines  src/Entity/DailyStats.php
class DailyStats
{
    /**
     * @Groups({"daily-stats:read"})
     */
    public ;
}
Whenever API Platform sees an embedded object:

 61 lines  src/Entity/DailyStats.php
class DailyStats
{
    /**
     * The 5 most popular cheese listings from this date!
     *
     * @Groups({"daily-stats:read"})
     */
    public $mostPopularListings;
}
It looks at that object - so CheeseListing - and looks to see if any of its fields are in that same group. And there are no properties inside CheeseListing that have the daily-stats:read group:

 219 lines  src/Entity/CheeseListing.php
class CheeseListing
{
    /**
     * @Groups({"cheese:read", "cheese:write", "user:read", "user:write"})
     */
    private $title;
    /**
     * @Groups({"cheese:read"})
     */
    private $description;
    /**
     * @Groups({"cheese:read", "cheese:write", "user:read", "user:write"})
     */
    private $price;
    /**
     * @Groups({cheese:write})
     */
    private $isPublished = false;
    /**
     * @Groups({"cheese:read", "cheese:collection:post"})
     */
    private $owner;
}
So... it makes sense that when we embed CheeseListing into DailyStats, none of its fields are included.

But, hold on a minute. Usually when API Platform sees an embedded object... and sees that no fields will be exposed on that object, instead of embedding it, it uses its IRI.
So the question is: why is it not doing that? Why do we have an array of embedded objects instead of an array of IRI strings?