19.2. Why/When a Many Relation is IRI Strings vs Embedded
API Platform Doesn't know Your Collection Type

So the question is:  Why do we have an array of embedded objects instead of an array of IRI strings?

     * @Groups({"daily-stats:read"})
     */
    public $mostPopularListings;
}
From API Platform's perspective, it knows that this will hold an array, but it does not know that it will hold an array of CheeseListing objects
And that changes how it behaves.
Ok, in truth, this CheeseListing[] doc should probably be enough to tell API Platform that this is an array of CheeseListing objects
But for some reason it doesn't read that syntax.
It's not important, but let's move the documentation to the property directly. Say @var array then greater than, less than with CheeseListing inside
This syntax is... maybe a bit lesser-known than the CheeseListing[] syntax. It comes from PSR-5 - a draft standard for PHP documentation.
As soon as we make this change... when we refresh... check it out! Suddenly it is the array of IRI strings that we expected! If we added the daily-stats:read group to any of the properties in CheeseListing, it would become an embedded object, but since we haven't, we get the expected IRI strings.
Now... the only problem with the array<CheeseListing> syntax is that... PhpStorm doesn't understand it. If we say $this->mostPopularListings[0]->... we get zero auto-completion.
The workaround is to use both formats: this is an array of CheeseListing or CheeseListing[]
And now PhpStorm is happy. This isn't ideal, but hopefully we won't need both formats in the future.
But... hold on. We just added documentation that changed the way API Platform serialized our object. Why was that needed? Can't API Platform figure out that this is an array of CheeseListing objects when we... pass it an array of CheeseListing objects? Why was the extra documentation even needed?
Let's dive deeper into that question next.