20.1. Collection "Types" and readableLink
Something a bit odd just happened: in order for API Platform to correctly serialize the mostPopularListings collection, we had to explicitly tell it what was inside the collection. Why?
To learn what's going on, let's look at another example. Inside User, we have a cheeseListings property, which is writable in our API, but isn't readable:

 288 lines  src/Entity/User.php
class User implements UserInterface
{
    /**
     * @Groups({"user:write"})
     */
    private $cheeseListings;
}
There is also a getPublishedCheeseListings() method, which is part of the API and we actually gave it the cheeseListings name:

 288 lines  src/Entity/User.php
class User implements UserInterface
{
    /**
     * @Groups({"user:read"})
     * @SerializedName("cheeseListings")
     */
    public function getPublishedCheeseListings(): Collection
    {
    }
}
Let's put in our lab coats and do an experiment! Science! Start by removing the SerializedName annotation
We're still going to expose this method, but it will use its natural name: publishedCheeseListings. Then, up on the cheeseListings property add user:read to also expose this
Let's see what it looks like! Head over to /api/users.jsonld and... cool! Each User now has cheeseListings and publishedCheeseListings properties and they're both embedded objects. The reason why is that the $title and $price properties in CheeseListing have the user:read group.