21.1. Custom Resource GET Item
Our GET collection operation for DailyStats is working nicely. But you know what? I'd love to also be able to fetch stats for a single day. The docs say that this operation already exists... but we know it's a lie!
If you look at the top of the DailyStats class, we kind of added the get item operation... but we made it return a 404 response
We did that as a workaround so that API Platform could generate an IRI for DailyStats. Now I want to make this truly work.
Remove all of the custom config so that we now have a normal get item operation and a normal get collection operation
We did that as a workaround so that API Platform could generate an IRI for DailyStats. Now I want to make this truly work.
Remove all of the custom config so that we now have a normal get item operation and a normal get collection operation
To get an item operation working, we need an item data provider. Since we don't have one yet, 404!