21.2. Custom Resource GET Item
Adding ItemDataProviderInterface
No problem for us: we've done this before! Add another interface called ItemDataProviderInterface
Then, down here, go to "Code"->"Generate" - or Command+N on a Mac - select "Implement Methods" and implement the getItem() function that we need
Our job here is to read this $id and return the DailyStats object for that $id, or null if there is none.

"Dynamic" Data from a JSON File
Before we do that, let's make all of this a bit more realistic. Our collection operation returns some hardcoded DailyStats. Let's pretend that we have a JSON file filled with this data... or maybe in a real project, you might have an external API you could talk to in order to fetch the data.
If you downloaded the course code, then you should have a tutorial/ directory with a fake_stats.json file inside. This will be our data source. Right next to this is a StatsHelper class, which holds code to read from that file.
Copy both of these. Then create a new directory in src/ called Service/ and paste them both there
Perfect. Let's take a quick look at the new StatsHelper
There's nothing fancy: it has three public methods - fetchMany(), where you can pass it a limit, offset and even some filtering criteria, which we'll talk about later - fetchOne(), where you pass a date string and also a count() method.
And... that's basically it. The rest of this file is boring code to read that fake_stats.json file, parse through it and create DailyStats objects.