21.3. Custom Resource GET Item
Using the "Dynamic" Data
In DailyStatsProvider let's use this! We won't need the CheeseListingRepository anymore: StatsHelper takes care of all of that. So autowire it instead: StatsHelper $statsHelper, then $this->statsHelper = $statsHelper and rename the property to $statsHelper
We can also get rid of couple of use statements
Down in getCollection(), it's now as simple as return $this->statsHelper->fetchMany(). For now, pass it no arguments.
Cool! Let's see if that works. Go back to the collection endpoint, refresh and... yes! We get a big list of DailyStats data coming from that JSON file!

Finishing getItem()
Let's use StatsHelper to finish getItem(). Thanks to the supports() method, our getItem() method should be called every time a request is made to an "item" operation for DailyStats
Let's make sure that's working with dd($id)
Back at the browser, go forward, refresh and... nice! Our date string is dumped.