22.1. Custom Paginator
When an API resource is a Doctrine entity like CheeseListing, life is good! We get tons of things for free, like pagination, access to a bunch of pre-built filters and one time I swear I got free pizza.
But... these filters will not work for custom resources... because API platform has no idea how to work with your custom data source. Nope, if you want filters or pagination on a custom API resource, you need to some work yourself.
But no worries: it's not too hard and we're kicking butt anyways. Let's start with pagination.

Pagination and your Data Provider
Right now, our collection endpoint lists every daily stats items... no matter how many there are.
Here's how pagination works: inside our data provider, instead of returning an array of DailyStats objects from getCollection(), we're going to return a Paginator object that contains the DailyStats objects that should be returned for whatever page we're on
In the src/DataProvider/ directory, create a new PHP class called DailyStatsPaginator. The only rule is that this needs to implement a PaginatorInterface from ApiPlatform
I'll go to "Code"->"Generate" - or Command+N on a Mac - and select "Implement Methods". Select all five methods we need.
Perfect! Oh, but I'll move count() to the bottom... it just feels better for me down there
To get this working, let's return some dummy data! In getLastPage(), pretend that there are two pages, for getTotalItems(), pretend there are 25 items, for getCurrentPage(), pretend we're on page 1 and for getItemsPerPage() return that we want to show 10 items per page. For count() return $this->getTotalItems()
Oh, yea know what? When I recorded this, I made a careless mistake. The count() method should actually return the number of items on this page - not the total number of items. I'll mention that in a few minutes when we actually have items to count.