22.2. Custom Paginator
Ok: there are no DailyStats objects inside this class yet... but let's do as little as possible to see if we can get things working.
Back in the provider, instead of returning the array of objects, return a new DailyStatsPaginator
Ooh. Let's try it. Head back over and refresh the collection endpoint. And... error!

  Class DailyStatsPaginator must implement the interface Traversable as part of either Iterator or IteratorAggregate.

Wow. Okay. So the PaginatorInterface gives API Platform a bunch of info about pagination, like how many pages there are and what page we're currently on. But ultimately, whatever we return from getCollection() needs to be something that API platform can loop over - can iterate over.

In other words, we need to make our paginator iterable. One way to do that is to add a second interface: \IteratorAggregate
And then, I'm going to create a new private property called $dailyStatsIterator
Finally, thanks to the new interface, at the bottom, go to "Code"->"Generate" - or Command+N on a Mac - and select "Implement Methods". This requires one new function called getIterator()
Inside, first check to see if $this->dailyStatsIterator === null. If it is null, set it: $this->dailyStatsIterator = and use another core class new \ArrayIterator(). For now, pass this an empty array. I'll put a little TODO up here:

  todo - actually go "load" the stats
At the bottom of the method, return $this->dailyStatsIterator
Basically, we're creating an iterator with the DailyStats objects inside. Well, it's an empty array now, but it will hold DailyStats soon. The property and if statement are there in case getIterator() is called multiple times. If it is, this prevents us from unnecessarily creating the iterator again.
Anyways, thanks to this, our paginator class is now iterable! And when we refresh the collection endpoint... it works! I mean, there's nothing inside the collection - hydra:member is empty - but you can see totalItems 25 and links below to tell us how to get to the first, next and last pages.