22.3. Custom Paginator
Iterating over DailyStats in the Paginator
To do the heavy lifting of loading the DailyStats objects, we can leverage the StatsHelper object, which we have access to inside of DailyStatsProvider.
Inside DailyStatsPaginator, add a constructor: public function __construct() with StatsHelper $statsHelper
Now, this class is not a service... So Symfony is not going to autowire this argument. This is really a "model" class that represents a paginated collection of DailyStats. But in a minute, we'll pass StatsHelper directly when we create this object.
Anyways, hit Alt+Enter and go to "Initialize properties" to create that property and set it
Then, before we use it, go to DailyStatsProvider and pass $this->statsHelper when we instantiate DailyStatsPaginator
For now, we're going to completely ignore pagination and still show all of the DailyStats no matter which page we're on. To do that, down in getIterator(), instead of the empty array, pass $this->statsHelper->fetchMany()
Oh, and by the way, you could now update the count() method to leverage getIterator(): return iterator_count($this->getIterator())
Anyways, let's check it out! Head over to your browser, refresh and... it works! Well, hydra:member still contains every DailyStats record... but all the pagination info is there.

So next, let's finish this! To do that, we'll need to ask API Platform:

  Hey! What page are we on?

And also:

  Yo! How many items per page should I show?

Because technically, that's configurable via annotations.