23.1. Pagination Context
We... sort of have pagination working? Except that the number of pages and items per page are hard coded. Oh, and we're always returning every DailyStats, no matter which page we're on. Ok, so it's not really working, but we're close.
There are two things that we need to know inside of DailyStatsPaginator: we need to know what the current page is - which is normally in the URL - and the max results per page that we should be showing. If we had those two pieces of info, we could fill in all of the methods.
Since these are, sort of, "options" that control the behavior of our paginator, let's force them to be passed in via the constructor. Add int $currentPage and int $maxResults arguments to the constructor. I'll even add a comma to separate them!
Next, hit Alt+Enter and go to "Initialize Properties" to create those two properties and set them
Sweet! Now in getCurrentPage(), return $this->currentPage and in getItemsPerPage(), return $this->maxResults

Completing all the Paginator Methods
And... we even have enough info to complete the other methods. For getLastPage(), I'll paste in a calculation
This returns the ceiling of the total items divided by the max items per page. And if that equals zero because there are no results, return 1.
Next, in getTotalItems(), return $this->statsHelper->count()
That method returns the total number of results, not just the results on this page.
Finally, in getIterator(), we need to figure out which results we should show based on the current page. We'll do that by calculating a limit and offset. Say $offset = and I'll paste in another calculation
The offset is the current page minus one, times the items per page.
Now, for fetchMany(), this accepts limit and offset arguments. Pass it the limit - $this->getItemsPerPage() - and then $offset
Phew! Our paginator is now 100% ready. To test it, open DailyStatsProvider. We now need to pass it the current page and the max items per page. To start, let's hardcode these: pretend we're on page 1 and we want to show 3 items per page so that pagination is really obvious
Let's see what it look like! Refresh the page and... awesome! 3 results, 30 totalItems - which is correct - and we're currently on page 1, next is page 2 and it will take us 10 pages to get through all 30 results. Our paginator is alive!