23.2. Pagination Context
The Pagination Object
All we need to do now is remove these hard-coded values. So how do we figure out what the current page is... or the max items per page that we should be showing? We could just choose any number we want and put it here for the max items. And in a project, that would be fine.
But technically, the max items per page is something that is configurable via the ApiResource annotation. And even the query parameter that's used for pagination can be changed - it doesn't need to be ?page=.
My point is: we don't need to hardcode the max per page or read the page query parameter directly because API Platform already has this info! Where? It's hiding in a service called Pagination: a service that we can autowire.
Add a second argument to DailyStatsProvider: Pagination - the one from DataProvider and call it $pagination
I'll hit Alt+Enter and go to Initialize Properties to create that property and set it
Ok: Pagination is an object that knows everything about the current pagination situation. So it has methods like getPage() and getOffset(), which is calculated from the current page and max items per page. We're going to use a - kind of strange - method called getPagination() which returns 3 pieces of info as an array.
Check it out: use the odd function list() to create three variables - $page, $offset and $limit - and set this to $this->pagination->getPagination() passing $resourceClass and $operationName
Notice that there is a third argument - $context. I'm not going to pass that simply because I don't have $context in this method. But if you did want to support the full features of the pagination system - there are a few edge cases where pagination changes based on the context - then make your class implement ContextAwareCollectionDataProviderInterface, which allows you to have a $context argument on getCollection().
Anyways, hold Command or Ctrl to jump into the getPagination() method. This returns an array containing the current page, the offset and the limit. We're using the strange list() function as a quick way to create three new variables:  set to the array's zero index, $offset set to the 1 index and $limit to the 2 index.
Thanks to this, below, we can use $page and $limit
We don't actually need $offset because we're calculating that ourselves.
Let's see if it works! Move over, refresh and this time... hmm. It looks like it's not paginating at all. The problem is that API Platform allows 30 items per page by default. And our JSON file has... yep - 30 items in it.